requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.2
Django==1.11.6
python-dateutil==2.6.1
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.1
ephem==3.7.6.0
idna==2.6
jwcrypto==0.4.2
kombu==4.1.0
pingpp==2.2.0
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.7.11
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
xmltodict==0.11.0
